Adding RFM95 class
