+++ platformio.ini
@@ -21,17 +21,23 @@ lib_deps =
 [env:F32u4_RFM95_MCP9808]
 board = feather32u4
 lib_deps = 
+    ${env.lib_deps}
     Adafruit MCP9808 Library
     Wire
-src_filter = +<LoRa/MCP9808> -<LoRa/Gateway> -<RFM69/>
+src_filter = +<LoRa/MCP9808> +<LoRa/> -<LoRa/Gateway> -<RFM69/>
+upload_port = COM13
 
 [env:F32u4_RFM95_Gateway]
 board = feather32u4
-src_filter = +<LoRa/Gateway> -<LoRa/MCP9808> -<RFM69/>
+lib_deps = 
+    ${env.lib_deps}
+src_filter = +<LoRa/Gateway> +<LoRa/> -<LoRa/MCP9808> -<RFM69/>
+upload_port = COM26
 
 [env:F32u4_RFM69_MCP9808]
 board = feather32u4
 lib_deps = 
+    ${env.lib_deps}
     Adafruit MCP9808 Library
     Wire
 src_filter = +<RFM69/MCP9808> -<RFM69/Gateway> -<RFM69/DHT11> -<LoRa/> 
@@ -39,16 +45,20 @@ src_filter = +<RFM69/MCP9808> -<RFM69/Gateway> -<RFM69/DHT11> -<LoRa/>
 [env:Uno_RFM69_DHT11]
 board = uno
 lib_deps = 
+    ${env.lib_deps}
     SimpleDHT
 src_filter = +<RFM69/DHT11> -<RFM69/Gateway> -<RFM69/MCP9808> -<LoRa/> 
 
 [env:Uno_RFM69_MCP9808]
 board = uno
 lib_deps = 
+    ${env.lib_deps}
     Adafruit MCP9808 Library
     Wire
 src_filter = +<RFM69/MCP9808> -<RFM69/Gateway> -<RFM69/DHT11> -<LoRa/> 
 
 [env:Uno_RFM69_Gateway]
 board = uno
+lib_deps = 
+    ${env.lib_deps}
 src_filter = +<RFM69/Gateway> -<RFM69/MCP9808> -<RFM69/DHT11> -<LoRa/>
